Change `Entity` and other generational IDs to use `u64` instead of `u32`

+++ crates/vizia_core/src/binding/map.rs
@@ -4,6 +4,6 @@ use vizia_id::{
 };
 
 #[derive(Clone, Copy, PartialEq, Eq, Hash)]
-pub struct MapId(u32);
+pub struct MapId(u64);
 
 impl_generational_id!(MapId);
